requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
pydantic==2.10.6
anthropic==0.49.0
python-dotenv==1.1.0
